handle unsafe filenames in dump_files_in_dir (SOFTWARE-1773)
Safety first!

This function was incredibly unsafe.  A carefully named file in any
directory that gets dumped with this function would allow arbitrary bash
code injection.  Fortunately, this wasn't that big of a deal, since the
directories that get dumped are all root-owned and only root-writable, so
only root on the system could have named such files.

Still, any files with whitespace or shell special characters would not get
handled properly.

("sort -z", which is available on EL5+, is added here to handle filenames
with embedded newlines, but frankly that wasn't the worst offender in the
original function.)